[AUTOCLAUDE] split things up
iteratively rewrite the files in src/ to make it more readable and pedagogic. splitting complex things out
  etc. the final result should not change any behavior, but it should be easier to follow the logic.

Total cost: .86
Total duration (API): 8m 43.7s
